Add test for backlogged auto-reload timer (#553)
* Add test for backlogged auto-reset timer

Must call vTimerDemoIncludeBacklogTests() to activate.

* Fix little style issues

- Update to new header
- Add parentheses (style)
- Remove trailing space accidentally introduced

* Don't verify full clearing of backlog

The timer task should be free to process the stop request without a
specific requirement to work through the backlog of callbacks for the
stopped timer.  The timer task should be allowed to cancel the entire
backlog, part of it, or none of it.  In other words, once the
application sends the stop request, it should not depend on receiving
any more callbacks, even backlogged ones.

The only requirement here is that the timer stop as requested.

Co-authored-by: Joseph Julicher <jjulicher@mac.com>
Co-authored-by: Cobus van Eeden <35851496+cobusve@users.noreply.github.com>